version 86.7
Finally finish incorperating index-based part structure with document-updating behavior. The final result breaks up the monolithic part class and makes a hierarchy of types just using native js classes.

The surprising nature of the user interaction design (involving an anchored selection so that one may click a first item, hold shift, and click a second to select a line of items) had to be realized in the methamatics: the cardinality of that configuration space is n2^(n-1)+1 rather than a naive n2^n which would have allowed cases where the starting point of selection is a deselected item - approximately half of the n2^n cases.

indexOf behavior is slapped in and untested but unneeded in the working demo.