0.89.0
Ready for GitHub pages, the site has finally been removed off of PHP. At long last, the PWA can be served from simple platforms like GitHub pages. There is even an opportunity to serve the platform without a serviceWorker, albiet not as a PWA.

There is still a bug related to the dev. prefix overlapping with the non-prefixed github.io domain. Solution is probably to make github pages link to the release version of the application only.