0.90.0
The default charset of the whole server is now utf-8 (yay!). Might need a tiny bit of attention to deploy on a non-appache server.

The entire boot flow has been reworked.

We have now removed the populate function and renamed enter and leave to create and destroy. The populate function was only being used reliably by decision and composite and was introducing a liability in regards to upstream vs downstream updating (child notify was not trigging parent populate but parent populate was trigging child populate).

Populate was being used as both a down-only rendering callback and as down propagation logic. Now, a new method called checkout has been used which only applies to decisions and composites and only handles down propagation with no rendering.

Rendering work that was in the populate of a part has been moved to the create and destroy of it's subpart and as designers we ensure all parts which can do "checkout" have the right corresponding subparts.
---
The web app at kireji.io offers selecting between different grammar models.

basic-001: nv
basic-002: anv
basic-003: (1 + a + a^2 + a^3)nv

This concept can be expanded and reused extensively.