0.90.1
Because github doesn't have a staging subdomain, this commit tries to associate it with the staging deployment instead of the live deployment. The ultimate test if it worked, however, will be when we "deploy to github"

Continue to lazily fix some ui rendering bugs on kireji. This is also a process of me learning to correctly work within my own constraints without taking shortcuts. There is an easy way to develop applications on this platform and then there is a right way. Kireji.io is currently the testing ground to find out which is which.