0.100.0
Rename Framework.archive to Framework.DNSRoot.

All new propagation flow. Now, the numerical state and document state propagate separately. The state is propagated firs. If we are not on the primary layer, propagation is done. If we are on the primary layer, the document is updated after state propagation.

The new flow allows the document update process to benifit from complete knowledge of the state the document is going to be in after the update is complete. This fixes annoyances where the state information of tree leaves was outdated during document update because the document update and state change events used to propagate leafward at the same time even though rootward nodes need to perform document updates before leafward ones.

Adds animated buttons to orenjinari.com.

Conjunctions and disjunctions now have the ability to call super with no arguments, so that type initailization arithmetic can be defered for types which require access to `this` in order to dynamically assign themselves factors and addends respectively.