0.115.54
# Continue Refactor

## Continue improving cross-site transition performance.

Fix incorrect play order for the menu caused by the move to alphabetical order. Future idea: feature a method of overriding the alphabetical subpat order. It doesn't have to be stored literally, it can be ignored by build packing and used merely as a guide for the correct order in which to define properties on the later-serialized object.