0.118.0
# Minor Refactor

This is a minor refactor with five major objectives:
1. Keep all property defining calls in either "build*.js" scripts or "*install.js" scripts. This is working towards a "preventExtensions" option which will let us lock down parts after the initial hydration, making it harder to accidentally assign enumerable properties to them.
2. Move the task menu from the desktop's root level to inside the task bar. It is, after all, only there as part of the task bar. Clean up some old code along the way.
3. Remove the old fetchSync function, clean up old service worker hand-off features and adjust the flow of onfetch in the service worker to align propertly with the current nginx and node.js file fetching behavior. TODO: Consider an abstraction which unites the node.js server with the serviceworker onfetch processing.
4. Remove the "theme" concept and replace it with the "root application" concept. This moves the theme pseudo-arm and themes set from the desktop root to the DNS root and renames all "theme" instances to "application". Now, the backend will include the theme as an aspect of the host upon fulfilling requests, removing the need for a default theme and complex theme switching messages.
5. Enable localhost applicaton testing again. This is critical for pre-production integration testing. Here, the protocol is maintained as http (which allows us to skip any local security certificate checks while still obtaining a secure context) and all application names are appended to ".localhost:3000" to ensure they can all be integration tested in the browser with the same cross-origin interactions as in production.

It also re-adds the transformation from subsomain name to valid javascript identifier (kabak-case to camel-case).

All const-*.js files have been removed. Only the behavior of const-method.js was retained (now called simply constants.js).

The window environment was renamed the "client" environment so that the "hydration part" - which sets the state of the O/S using the addressbar URL - can be renamed "client."

This creates a consistency across three of the environments: all three have a same-named facet which bears the responsibility of setting the URL-based route for the given environment. The only other place where that route is set is during the validation phase of the server build process.