0.118.4
Vastly improve image rendering flow. New flow is pretty slick in testing. The `part.render()` function now processes all internal png and gif rendering so that it provides only a transparent GIF of the right width and height as the true image by extracting a few bytes from the right location in the base64 data string. The resulting GIFs are quite small (always 42 bytes) and provide a placeholder that is safe to inline even many times on the page without losing out on the structural information that having the correct width and height of the image and having a true `<img>` tag can provide.

The result is a small (15-50kb) server-rendered HTML snippet for virtually all O/S states which can deliver a highly accurate skeleton of the page, sans the true image. A bit of data is added to the base64 image header to remind us later of the image's true owner and location, as per the part archive. At the same time, the server-rendered page has an empty inline stylesheet called #img-css which is ready to be populated once that archive is on the client in the form of the service worker.

As soon with the HTML is parsed and the service worker is rendered, we are able to insert css which has exactly one copy of each image as a `url(data:image/...)` variable which populates the images based on their src tag. The result is that the hydration process quickly shows the real images.

After that initial fetch process, the service worker takes over. Now, the above css snippet is _inlined_ in the service worker-provided HTML, which makes the image rendering _instant_ after the very first fetch. Even when there are service worker updates, the server should never again have to show that initial blink.

Refreshes and cross-origin travel is seamless (once the destination origin has been visited at least once), with very tiny fetch sizes from the server and with less processing time than if we inlined each of the base64 images in `<img src="...">` directly.