0.119.0
Reduce precision of scroller from 100,000 to 10,000. This helps reduce entropy and can be changed easily at any time. However, changing this is API-breaking, resulting in a major change.

Develop www.ejaugust.com. Add color mode-relative accent colors. Improve layout and interaction. Add client engagement to the service worker update process. While not active yet, this method will result in a future where there can be a notification or popup which will allow the user to optionally update the service worker immediately or have it update the next time they open the app.

Add vertical scroller to explorer.www.core.parts. The sticky heading is not staying there during the initial load, resulting in a FOUC.

Fix some missing images on www.orenjinari.com. Still unable to fix a FOUC only happening to the background to the wallpaper scroller - not sure why. Could be an issue specifically with the scroller, and could be related to the issue with the www.core.parts scroller.