0.120.1
Don't automatically install service worker. This prevents a few issues like infinite looping installs in the case where the user has "update on refresh" activated in dev tools and the occasional loss of the interaction that took place after the update was found but before the refresh.

When not in production, popup to confirm instant update versus production-like delayed update behavior.