0.124.0
Disabled native scrollbars and implemented custom scrollbars. This is the only way to achieve the vintage look accurately, and gives us complete control over the scrollbar from here on out. It required significant restructuring to scroll containers.

Added new abstract parts for top-level domains and apex domains. Added new part icons for the DNS Root, top-level domains, and apex domains.

Added an event system to parts. Parts can now add and remove event listeners to other parts. This is perfect for dynamic connections. Currently, this is used to update the selected part panel in www.kireji.app whenever said part experiences a population of its own.

Added a base description and title system which allows abstract parts to set these properties in a way that won't be inherited by child elements.

Improved some titles and descriptions.

Layout changes to the part editor at www.kireji.app. Also, all parts (including abstract parts) are now available in the editor. A link where the host of the selected part's prototype is shown allows users to quickly select said prototype part, allowing them to follow the chain down to the base part. Add a sidebar toggling method and corresponding button for the editor panel.

Fixes to the hot-key manager (which was checking the wrong part for determining if the user agent is on mac) and improved integration with native back/forward usage. Each history entry in the native browser presents an undoable event. Each part which wants to have its interaction recorded as a history event must do so manually by calling pushState, as the defauly behavior for all routeID changes at the root level remains replaceState. Cross-origin navigation _always_ pushes a state, and this is an unavoidable security feature of the browser.

Add an html superspript markup option for `scientific()` (previously, it only had the option which remains the default option: using unicode superscript characters, which is still ideal for console logging).

A handful of non-blocking visual bugs remains on this version and should be improved in the next version.