0.125.3
Add descriptions to each of the o/s facets.

Improved angle to line on open box flaps for the facet.core.parts type icon.

Add a version-based scope for the PWA manifest.json. This allows users to install multiple versions of the application on their device at the same time and prevents clashes and overriding when a new version of the app is available.

Add commented-out hint for future, which may allow us to change the PWA title bar color based on the color mode. This will likely need to be integrated with color.desktop.parts in order to work properly.

Delete unused 1x1.gif placeholder.