0.127.4
## Hot Key Manager Overhaul
Significant improvements to hot-keys.desktop.parts. Now, hot-keys can handle robust cross-origin keyboard sequences like undo and redo and we are using a cross-platform method that accounts well for macOS's unusual choice not to notify the browser of non-modifier keyup events for the duration that the user is holding the command key.

There is now a dedicated JSON table for linking keyboard shortcuts to methods on the hotKeys part itself. The table keys use a standardized notation in which the context key (platform agnostic abstraction of the Meta and Control keys) is always the first key in the list (if it should be pressed), the shift key is always the second (if it should be pressed), and there can be up to one non-modifier key at the end. These are delimited by the "+" character.

A new method at hotKeys.getComboString() returns the current combo string being pressed by the user.

## Notebook Edits
Fix typos on one of the quick tips in the notebook.

Improve the wording of the service worker update note in the notebook.