0.135.11
Remove logs now that menu highlighting bug is fixed. Bug was caused by a missing comma which caused a constant declared within the initial boot scope to be a global var. Then, other versions of the application with the same bug would apply over that global bar when they were imported as a module.