0.135.12
Stringify error log in server catch, to prevent second throw in case error thrown is not a string.

Get rid of the `part.render()` function and merge it with the worker onfetch handler since that's the only function that ever called it.

Get rid of the FileHeader class (yay!) since it was _mostly_ only used by the render function to discern the file MIME.

Implement the correct 501 status for error 501 pages.

Replace kireji framework logo on kireji.app with what used to be the core.parts icon, now increased to 192px so it can be an app icon.

Increase part.abstract.part's icon to 192px so that apps without icons can fallback to the abstract icon and still be a PWA.