0.137.0
# Heavy Refactor

Remove "www" subdomains and let apex domains be applications directly. This makes the hierarchy more shallow and clean. It removes the need to add an icon, title and description to the apex domain separately from the dedicated www domain that was on it. There was by definition a 1:1 correspondance between applications and apex domains. Now, all applications extend `application.abstract.parts`, (a mix) allowing us to abstract common application features and to know that all applications are mixes.