0.148.2
Fix inconsistency in vintage outset shading.

Make all EJ's Notebook notes mixes (for future tools and mini apps). Because the parent part ("notes") of all notes is a match, these mixes will not retain their state when navigating away to another note (or home). To retain a note's state when navigating away from the note, leave the note as a 1-state mix and allow it to display the state of another stateful part which is outsode of the "sections" part (similar to how the scroller is handled).