0.148.4
Fix problem that suprirsingly went unnoticed for some time: the "clean" step of the route event loop were not doing anything at all. The property `part.dirty` was being deleted but they are not modifiable properties and cannot be deleted. Attempting to delete it produces no error but does not unset the dirty flag flag from `true`.

Now, it is set to false instead. This property is used as a gate to tell view population methods whether or not the part has undergone a route change since the last view update and is cleared after populating all views. It may not be necessary to use this flag at all, but some research needs to be done to determine that.